Update migration to fix error

+++ db/migrate/20220517204447_modify_contract_related_tables.rb
@@ -1,12 +1,18 @@
 class ModifyContractRelatedTables < ActiveRecord::Migration[6.1]
   def change
-    remove_column :contract_vehicles, :sins
+    if column_exists?  :contract_vehicles, :sins
+      remove_column :contract_vehicles, :sins
+    end
     add_column :contract_vehicles, :conctract_sub_category_id, :integer
     
-    remove_column :contract_vehicles, :contract_agency
+    if column_exists?  :contract_vehicles, :contract_agency
+      remove_column :contract_vehicles, :contract_agency
+    end
     add_column :contract_vehicles, :contract_agency_id, :integer
 
-    remove_column :contract_vehicles, :vehicle_type
+    if column_exists?  :contract_vehicles, :vehicle_type
+      remove_column :contract_vehicles, :vehicle_type
+    end
     add_column :contract_vehicles, :vehicle_type_id, :integer
     
   end
